extract routes
