+++ src/App.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+import { Provider } from 'react-redux';
+import store from './store/store';
+
+import Routes from './routes';
+
+const App = () => (
+  <Provider store={store}>
+    <Routes />
+  </Provider>
+);
+
+export default App;
