+++ src/components/About.js
@@ -1,9 +1,7 @@
 import React from "react";
 
-const About = () => {
-  return (
-    <h3>All information provide in this APP is not real time.</h3>
-  )
-}
+const About = () => (
+  <h3>All information provide in this APP is not real time.</h3>
+)
 
 export default About;
