+++ src/components/Navbar.js
@@ -1,13 +1,11 @@
 import React from "react";
 import { Link } from 'react-router-dom';
 
-const Navbar = () => {
-  return (
-    <nav className="flex">
-      <Link to="/">About</Link>
-      <Link to="/dashboard">DashBoard</Link>
-    </nav>
-  )
-}
+const Navbar = () => (
+  <nav className="flex">
+    <Link to="/">About</Link>
+    <Link to="/dashboard">DashBoard</Link>
+  </nav>
+)
 
 export default Navbar;
